Add refresh icon, move chart object to top level
