+++ stockflux-chart/src/App.js
@@ -2,10 +2,15 @@ import React, {useState} from 'react';
 import Chart from './components/Chart';
 import Components from 'stockflux-components';
 import { Intents } from 'stockflux-core';
-import { FaNewspaper, FaRegListAlt } from 'react-icons/fa';
+import { FaNewspaper, FaRegListAlt, FaSyncAlt } from 'react-icons/fa';
 import {InterApplicationBusHooks} from 'openfin-react-hooks';
+import bitflux from 'stockflux-bitflux/dist/bitflux';
 
 import './styles/app.css';
+
+const chart = bitflux.app();
+chart.periodsOfDataToFetch(1200);
+chart.proportionOfDataToDisplayByDefault(112/1200);
 
 const App = () => {
     const [symbol, setSymbol] = useState(null);
@@ -43,28 +48,39 @@ const App = () => {
         }
     }
 
+    const getData = (symbol) => {
+        if (symbol) {
+            chart.changeStockFluxProduct(symbol);
+        }
+    }
+
     return (
         <>
             <div className='main'>
                 <div className='main-content'>
                     <Components.Titlebar />
                     <div id="showcase-title">
-                        <div className="code">
-                            {symbol}
-                        </div> 
+                        {symbol && (
+                            <div className="code">
+                                {symbol}
+                            </div> 
+                        )}
                         <div className="name">
                             {name ? name : 'Generated Data'}
                         </div>
                         <div className="chart-nav-icons">
-                            <div className="chart-news-icon" onClick={onNewsClick}>
+                            <div className={"chart-nav-icon" + (symbol ? '' : ' icon-disabled')} onClick={onNewsClick}>
                                 <FaNewspaper />
                             </div>
-                            <div className="chart-watchlist-icon" onClick={onWatchlistClick}>
+                            <div className={"chart-nav-icon" + (symbol ? '' : ' icon-disabled')} onClick={onWatchlistClick}>
                                 <FaRegListAlt />
+                            </div>
+                            <div className={"chart-nav-icon" + (symbol ? '' : ' icon-disabled')} onClick={() => getData(symbol)}>
+                                <FaSyncAlt />
                             </div>
                         </div>
                     </div>
-                    <Chart symbol={symbol}/>
+                    <Chart getData={getData} chart={chart} symbol={symbol}/>
                 </div>
             </div>
         </>
